pyproject.toml
[tool.poetry]
name = "jwt"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.3"
djangorestframework-simplejwt = "^5.3.1"
djangorestframework = "^3.15.0"
django-restframework = "^0.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
